requirements.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
apex==0.1
appdirs==1.4.4
argcomplete==3.6.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
basicsr==1.4.2
beautifulsoup4==4.14.2
black==21.4b2
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.3.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.6.0
cuml-cu12==25.6.0
cupy-cuda12x==13.6.0
cuvs-cu12==25.6.1
cycler==0.12.1
Cython==3.0.2
dash==3.2.0
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu12==25.6.0
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
-e git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0#egg=detectron2
diffdist==0.1
diffusers==0.35.1
distributed==2025.5.0
distributed-ucxx-cu12==0.44.0
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.7.0
embreex==2.17.7.post6
evo==1.31.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.2
fastrlock==0.8.3
ffmpy==0.6.1
filelock==3.13.4
Flask==3.1.2
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2==2.1.2
gradio==5.17.1
gradio_client==1.7.1
groovy==0.1.2
grpcio==1.74.0
gsplat==1.5.3
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
icecream==2.1.8
idna==3.7
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
infinibatch==0.1.1
iopath==0.1.9
ipdb==0.13.13
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.3
jiter==0.11.1
joblib==1.5.2
json-tricks==3.17.3
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
libcudf-cu12==25.6.0
libcuml-cu12==25.6.0
libcuvs-cu12==25.6.1
libkvikio-cu12==25.6.0
libraft-cu12==25.6.0
librmm-cu12==25.6.0
libucx-cu12==1.18.1
libucxx-cu12==0.44.0
lightning-utilities==0.15.2
llvmlite==0.43.0
lmdb==1.7.5
locket==1.0.0
lxml==6.0.1
lz4==4.4.4
manifold3d==3.0.1
mapbox_earcut==1.0.3
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
mmcv==1.6.0
mmengine==0.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
MultiScaleDeformableAttention==1.0
mypy_extensions==1.1.0
narwhals==2.2.0
natsort==8.4.0
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numba==0.60.0
numba-cuda==0.11.0
numexpr==2.11.0
numpy==2.0.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.7
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.9.86
nvitop==1.5.3
nvtx==0.2.13
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.32.0
openai==2.7.1
opencv-python==4.10.0.84
orjson==3.11.3
packaging==25.0
pandas==2.2.3
parso==0.8.5
partd==1.4.2
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
platformdirs==4.4.0
plotly==6.3.0
plyfile==1.1.2
portalocker==3.2.0
progressbar==2.5
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocotools==2.0.10
pycollada==0.9.2
pydantic==2.12.4
pydantic_core==2.41.5
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pylibcudf-cu12==25.6.0
pylibraft-cu12==25.6.0
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
-e git+https://github.com/facebookresearch/pytorch3d.git@dd068703d1182256c7bacf3eb6014f24c1190369#egg=pytorch3d
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.1
pyzmq==27.1.0
qwen-vl-utils==0.0.11
raft-dask-cu12==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.1
referencing==0.36.2
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rich==14.0.0
rmm-cu12==25.6.0
rosbags==0.10.9
rpds-py==0.27.1
rtree==1.4.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.35.1
setproctitle==1.3.7
Shapely==1.8.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack_data==0.6.3
starlette==0.47.3
svg.path==6.3
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.17.0a20240519
tblib==3.1.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.12.0
timm==1.0.15
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
torch==2.5.0
torch-geometric==2.6.1
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.5.0
torchmetrics==1.8.1
torchvision==0.20.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
treelite==4.4.1
trimesh==4.7.4
triton==3.1.0
typeguard==4.4.2
typer==0.16.1
typer-slim==0.15.3
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.28
tzdata==2025.2
ucx-py-cu12==0.44.0
ucxx-cu12==0.44.0
urllib3==2.2.1
uvicorn==0.34.2
vhacdx==0.0.8.post2
viser==0.2.23
vision-datasets==0.2.2
wadler_lindig==0.1.5
wandb==0.21.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xformers==0.0.28.post2
